+++ backend-master/lib/database.rb
@@ -3,4 +3,3 @@
 require_relative 'database/base'
 require_relative 'database/in_memory_connector'
 require_relative 'database/mysql_connector'
-