+++ backend-master/lib/market/btc_market.rb
@@ -6,58 +6,81 @@ module PaymiumMarket
       BASE = 'BTC'
       QUOTE = 'EUR'
 
+      # a simplified order representation for easy manipulation
+      class SimpleOrder
+        extend Dry::Initializer
+
+        param :id
+        param :price
+        param :amount
+      end
+
       # sets the database connector
       def initialize(db_connector = PaymiumMarket::Database::MysqlConnector.new)
         @db = db_connector
       end
 
-      # creates a new order in the database
       def submit(order)
         @db.create(order)
       end
 
-      # gives the market price
       def market_price
-        (BigDecimal(max_bid) + BigDecimal(min_ask)) / 2.0
+        ((BigDecimal(max_bid.price) + BigDecimal(min_ask.price)) / 2.0).to_f
+      rescue NoMethodError
+        return 0
       end
 
-      # returns a hash with the market status
       def market_depth
         {
-          'bids' => bids,
+          'bids' => bids.values.sort_by { |order| - BigDecimal(order[0]) },
           'base' => BASE,
           'quote' => QUOTE,
-          'asks' => asks
+          'asks' => asks.values.sort_by { |order| - BigDecimal(order[0]) }
         }
       end
 
-      # removes an order from the database
       def cancel_order(id)
         @db.delete(id)
+      end
+
+      # simplified matching engine
+      def match
+        bid = max_bid
+        ask = min_ask
+        if bid.price == ask.price
+          begin
+            # change order status
+            # change users balance
+          rescue Error
+            #TODO
+          end
+          return 1 + match
+        end
+        0
       end
 
       private
 
       def bids
-        @db.bids.values.sort_by { |order| - order[0] }
+        @db.bids
       end
 
       def asks
-        @db.asks.values.sort_by { |order| - order[0] }
+        @db.asks
       end
 
       def max_bid
-        max_order = @db.bids.values.max_by { |order| order[0] }
-        return 0 if max_order.nil? || max_order[0].nil?
+        candidate = @db.bids.max_by {|k, v| BigDecimal(v[0])}
+        return SimpleOrder.new(-1, 0, 0) if candidate.nil?
 
-        max_order[0]
+        SimpleOrder.new candidate[0], candidate[1][0], candidate[1][1]
       end
 
       def min_ask
-        min_order = @db.asks.values.max_by { |order| order[0] }
-        return 0 if min_order.nil? || min_order[0].nil?
+        candidate = @db.asks.max_by {|k, v| BigDecimal(v[0])}
+        return SimpleOrder.new(-1, 0, 0) if candidate.nil?
 
-        min_order[0]
+        SimpleOrder.new candidate[0], candidate[1][0], candidate[1][1]
       end
     end
   end